fix the ereor

+++ pointers_arrays_strings/100-atoi.c
@@ -6,26 +6,26 @@
  */
 int _atoi(char *s)
 {
-	int si = 1, i = 0, res = 0, stop = 0;
+	long int si = 1, i = 0, res = 0, stop = 0;
 
 	while (s[i] != '\0')
 	{
-	    if (s[i] == '-' && s[i + 1] <= '9' && s[i + 1] >= '0')
-	    {
+		if (s[i] == '-' && s[i + 1] <= '9' && s[i + 1] >= '0')
+		{
 		    si = -1;
-	    }
-	    while (s[i] <= '9' && s[i] >= '0')
-	    {
+	    	}
+		while (s[i] <= '9' && s[i] >= '0')
+		{
 		    stop = 1;
 		    res = 10 * res + (s[i] - '0');
 		    i++;
-	    }
-	    if (stop == 1)
-	    {
+	    	}
+		if (stop == 1)
+	    	{
 		    break;
-	    }
-	    i++;
-    }
+	    	}
+	    	i++;
+    	}
 return (res * si);
 }
